+++ src/main.rs
@@ -1,14 +1,10 @@
 use clap::Parser;
-use core::str;
 use std::num::NonZeroUsize;
 
 use anyhow::Context;
-use arrow2::{array::Float64Array, chunk::Chunk, datatypes::Field};
 
 use std::{fs::File, io};
 
-use num_format::{Locale, ToFormattedString};
-use serde_json::json;
 use tracing::info;
 use tracing_subscriber::{filter::LevelFilter, fmt, prelude::*, EnvFilter};
 
@@ -17,6 +13,7 @@ use noodles_bgzf as bgzf;
 
 mod alignment_parser;
 mod bootstrap;
+mod bulk;
 mod em;
 mod prog_opts;
 mod single_cell;
@@ -24,11 +21,7 @@ mod util;
 
 use crate::prog_opts::{Args, FilterGroup};
 use crate::util::normalize_probability::normalize_read_probs;
-use crate::util::oarfish_types::{
-    AlignmentFilters, EMInfo, InMemoryAlignmentStore, TranscriptInfo,
-};
-use crate::util::read_function::read_short_quant_vec;
-use crate::util::write_function::{write_infrep_file, write_output};
+use crate::util::oarfish_types::{AlignmentFilters, TranscriptInfo};
 use crate::util::{binomial_probability::binomial_continuous_prob, kde_utils};
 
 fn get_filter_opts(args: &Args) -> AlignmentFilters {
@@ -74,36 +67,6 @@ fn get_filter_opts(args: &Args) -> AlignmentFilters {
     }
 }
 
-/// Produce a [serde_json::Value] that encodes the relevant arguments and
-/// parameters of the run that we wish to record to file. Ultimately, this
-/// will be written to the corresponding `meta_info.json` file for this run.
-fn get_json_info(args: &Args, emi: &EMInfo, seqcol_digest: &str) -> serde_json::Value {
-    let prob = if args.model_coverage {
-        "scaled_binomial"
-    } else {
-        "no_coverage"
-    };
-
-    json!({
-        "prob_model" : prob,
-        "bin_width" : args.bin_width,
-        "filter_options" : &emi.eq_map.filter_opts,
-        "discard_table" : &emi.eq_map.discard_table,
-        "alignments": &args.alignments,
-        "output": &args.output,
-        "verbose": &args.verbose,
-        "single_cell": &args.single_cell,
-        "quiet": &args.quiet,
-        "em_max_iter": &args.max_em_iter,
-        "em_convergence_thresh": &args.convergence_thresh,
-        "threads": &args.threads,
-        "filter_group": &args.filter_group,
-        "short_quant": &args.short_quant,
-        "num_bootstraps": &args.num_bootstraps,
-        "seqcol_digest": seqcol_digest
-    })
-}
-
 fn main() -> anyhow::Result<()> {
     let env_filter = EnvFilter::builder()
         .with_default_directive(LevelFilter::INFO.into())
@@ -132,7 +95,6 @@ fn main() -> anyhow::Result<()> {
 
     let filter_opts = get_filter_opts(&args);
 
-    let flen = std::fs::metadata(&args.alignments)?.len();
     let afile = File::open(&args.alignments)?;
 
     let worker_count = NonZeroUsize::new(1.max(args.threads.saturating_sub(1)))
@@ -193,7 +155,7 @@ fn main() -> anyhow::Result<()> {
                     .add_directive("oarfish::single_cell=warn".parse().unwrap())
             } else if args.verbose {
                 EnvFilter::new("TRACE")
-                    .add_directive("oarfish=warn".parse().unwrap())
+                    .add_directive("oarfish=info".parse().unwrap())
                     .add_directive("oarfish::single_cell=trace".parse().unwrap())
             } else {
                 EnvFilter::new("INFO")
@@ -203,7 +165,6 @@ fn main() -> anyhow::Result<()> {
         })?;
 
         single_cell::quantify_single_cell_from_collated_bam(
-            flen,
             &header,
             &filter_opts,
             &mut reader,
@@ -212,110 +173,15 @@ fn main() -> anyhow::Result<()> {
             seqcol_digest,
         )?;
     } else {
-        // now parse the actual alignments for the reads and store the results
-        // in our in-memory stor
-        let mut store = InMemoryAlignmentStore::new(filter_opts, &header);
-        alignment_parser::parse_alignments(&mut store, &header, &mut reader, &mut txps)?;
-
-        // print discard table information in which the user might be interested.
-        info!("\ndiscard_table: \n{}\n", store.discard_table.to_table());
-
-        // no longer need the reader
-        drop(reader);
-
-        // if we are using the KDE, create that here.
-        let kde_opt: Option<kders::kde::KDEModel> = if args.use_kde {
-            Some(kde_utils::get_kde_model(&txps, &store)?)
-        } else {
-            None
-        };
-
-        if store.filter_opts.model_coverage {
-            //obtaining the Cumulative Distribution Function (CDF) for each transcript
-            binomial_continuous_prob(&mut txps, &args.bin_width, args.threads);
-            //Normalize the probabilities for the records of each read
-            normalize_read_probs(&mut store, &txps, &args.bin_width);
-        }
-
-        info!(
-            "Total number of alignment records : {}",
-            store.total_len().to_formatted_string(&Locale::en)
-        );
-        info!(
-            "number of aligned reads : {}",
-            store.num_aligned_reads().to_formatted_string(&Locale::en)
-        );
-        info!(
-            "number of unique alignments : {}",
-            store.unique_alignments().to_formatted_string(&Locale::en)
-        );
-
-        // if we are seeding the quantification estimates with short read
-        // abundances, then read those in here.
-        let init_abundances = args.short_quant.as_ref().map(|sr_path| {
-            read_short_quant_vec(sr_path, &txps_name).unwrap_or_else(|e| panic!("{}", e))
-        });
-
-        // wrap up all of the relevant information we need for estimation
-        // in an EMInfo struct and then call the EM algorithm.
-        let emi = EMInfo {
-            eq_map: &store,
-            txp_info: &txps,
-            max_iter: args.max_em_iter,
-            convergence_thresh: args.convergence_thresh,
-            init_abundances,
-            kde_model: kde_opt,
-        };
-
-        if args.use_kde {
-            /*
-            // run EM for model train iterations
-            let orig_iter = emi.max_iter;
-            emi.max_iter = 10;
-            let counts = em::em(&emi, args.threads);
-            // relearn the kde
-            let new_model =
-            kde_utils::refresh_kde_model(&txps, &store, &emi.kde_model.unwrap(), &counts);
-            info!("refreshed KDE model");
-            emi.kde_model = Some(new_model?);
-            emi.max_iter = orig_iter;
-            */
-        }
-
-        let counts = if args.threads > 4 {
-            em::em_par(&emi, args.threads)
-        } else {
-            em::em(&emi, args.threads)
-        };
-
-        let aux_txp_counts = crate::util::aux_counts::get_aux_counts(&store, &txps)?;
-
-        // prepare the JSON object we'll write
-        // to meta_info.json
-        let json_info = get_json_info(&args, &emi, &seqcol_digest);
-
-        // write the output
-        write_output(&args.output, json_info, &header, &counts, &aux_txp_counts)?;
-
-        // if the user requested bootstrap replicates,
-        // compute and write those out now.
-        if args.num_bootstraps > 0 {
-            let breps = em::bootstrap(&emi, args.num_bootstraps, args.threads);
-
-            let mut new_arrays = vec![];
-            let mut bs_fields = vec![];
-            for (i, b) in breps.into_iter().enumerate() {
-                let bs_array = Float64Array::from_vec(b);
-                bs_fields.push(Field::new(
-                    format!("bootstrap.{}", i),
-                    bs_array.data_type().clone(),
-                    false,
-                ));
-                new_arrays.push(bs_array.boxed());
-            }
-            let chunk = Chunk::new(new_arrays);
-            write_infrep_file(&args.output, bs_fields, chunk)?;
-        }
+        bulk::quantify_bulk_alignments_from_bam(
+            &header,
+            filter_opts,
+            &mut reader,
+            &mut txps,
+            &txps_name,
+            &args,
+            seqcol_digest,
+        )?;
     }
 
     Ok(())
